api/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.9
python-dotenv==1.1.1
gunicorn==22.0.0